[Tech] fix mission and dashboard table
