+++ backend/src/test/kotlin/fr/gouv/cacem/monitorenv/infrastructure/database/repositories/JpaDashboardRepositoryITest.kt
@@ -102,6 +102,6 @@ class JpaDashboardRepositoryITest : AbstractDBTests() {
         val dashboards = jpaDashboardRepository.findAll()
 
         // Then
-        assertThat(dashboards).hasSize(4)
+        assertThat(dashboards).hasSize(7)
     }
 }
